+++ pkg/studyengine/expressions_test.go
@@ -425,6 +425,158 @@ func TestEvalHasParticipantFlag(t *testing.T) {
 	})
 }
 
+func TestEvalResponseHasOnlyKeysOtherThan(t *testing.T) {
+	testEvalContext := EvalContext{
+		Event: types.StudyEvent{
+			Type: "SUBMIT",
+			Response: types.SurveyResponse{
+				Key:       "wwekly",
+				Responses: []types.SurveyItemResponse{},
+			},
+		},
+	}
+
+	t.Run("no survey item response found", func(t *testing.T) {
+		exp := types.Expression{Name: "responseHasOnlyKeysOtherThan", Data: []types.ExpressionArg{
+			{DType: "str", Str: "weekly.G1.Q1"},
+			{DType: "str", Str: "rg.mcg"},
+			{DType: "str", Str: "1"},
+			{DType: "str", Str: "2"},
+		}}
+		testEvalContext.Event.Response.Responses = []types.SurveyItemResponse{
+			{Key: "weekly.G1.Q2", Response: &types.ResponseItem{Key: "rg", Items: []types.ResponseItem{{Key: "mcg", Items: []types.ResponseItem{
+				{Key: "0"},
+			}}}}},
+		}
+		ret, err := ExpressionEval(exp, testEvalContext)
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+			return
+		}
+		if ret.(bool) {
+			t.Errorf("unexpected value: %b", ret)
+		}
+
+	})
+
+	t.Run("with response item found, but no response parent group", func(t *testing.T) {
+		exp := types.Expression{Name: "responseHasOnlyKeysOtherThan", Data: []types.ExpressionArg{
+			{DType: "str", Str: "weekly.G1.Q1"},
+			{DType: "str", Str: "rg.mcg"},
+			{DType: "str", Str: "1"},
+			{DType: "str", Str: "2"},
+		}}
+		testEvalContext.Event.Response.Responses = []types.SurveyItemResponse{
+			{Key: "weekly.G1.Q1", Response: &types.ResponseItem{Key: "rg", Items: []types.ResponseItem{{Key: "scg", Items: []types.ResponseItem{
+				{Key: "0"},
+			}}}}},
+		}
+		ret, err := ExpressionEval(exp, testEvalContext)
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+			return
+		}
+		if ret.(bool) {
+			t.Errorf("unexpected value: %b", ret)
+		}
+
+	})
+
+	t.Run("response group does include at least one", func(t *testing.T) {
+		exp := types.Expression{Name: "responseHasOnlyKeysOtherThan", Data: []types.ExpressionArg{
+			{DType: "str", Str: "weekly.G1.Q1"},
+			{DType: "str", Str: "rg.mcg"},
+			{DType: "str", Str: "1"},
+			{DType: "str", Str: "2"},
+		}}
+		testEvalContext.Event.Response.Responses = []types.SurveyItemResponse{
+			{Key: "weekly.G1.Q1", Response: &types.ResponseItem{Key: "rg", Items: []types.ResponseItem{{Key: "mcg", Items: []types.ResponseItem{
+				{Key: "0"},
+				{Key: "1"},
+				{Key: "3"},
+			}}}}},
+		}
+		ret, err := ExpressionEval(exp, testEvalContext)
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+			return
+		}
+		if ret.(bool) {
+			t.Errorf("unexpected value: %b", ret)
+		}
+
+	})
+
+	t.Run("response group is empty", func(t *testing.T) {
+		exp := types.Expression{Name: "responseHasOnlyKeysOtherThan", Data: []types.ExpressionArg{
+			{DType: "str", Str: "weekly.G1.Q1"},
+			{DType: "str", Str: "rg.mcg"},
+			{DType: "str", Str: "1"},
+			{DType: "str", Str: "2"},
+		}}
+		testEvalContext.Event.Response.Responses = []types.SurveyItemResponse{
+			{Key: "weekly.G1.Q1", Response: &types.ResponseItem{Key: "rg", Items: []types.ResponseItem{{Key: "mcg", Items: []types.ResponseItem{}}}}},
+		}
+		ret, err := ExpressionEval(exp, testEvalContext)
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+			return
+		}
+		if ret.(bool) {
+			t.Errorf("unexpected value: %b", ret)
+		}
+
+	})
+
+	t.Run("response group includes all and other responses", func(t *testing.T) {
+		exp := types.Expression{Name: "responseHasOnlyKeysOtherThan", Data: []types.ExpressionArg{
+			{DType: "str", Str: "weekly.G1.Q1"},
+			{DType: "str", Str: "rg.mcg"},
+			{DType: "str", Str: "1"},
+			{DType: "str", Str: "2"},
+		}}
+		testEvalContext.Event.Response.Responses = []types.SurveyItemResponse{
+			{Key: "weekly.G1.Q1", Response: &types.ResponseItem{Key: "rg", Items: []types.ResponseItem{{Key: "mcg", Items: []types.ResponseItem{
+				{Key: "0"},
+				{Key: "1"},
+				{Key: "2"},
+			}}}}},
+		}
+		ret, err := ExpressionEval(exp, testEvalContext)
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+			return
+		}
+		if ret.(bool) {
+			t.Errorf("unexpected value: %b", ret)
+		}
+
+	})
+
+	t.Run("response group includes none of the options", func(t *testing.T) {
+		exp := types.Expression{Name: "responseHasOnlyKeysOtherThan", Data: []types.ExpressionArg{
+			{DType: "str", Str: "weekly.G1.Q1"},
+			{DType: "str", Str: "rg.mcg"},
+			{DType: "str", Str: "1"},
+			{DType: "str", Str: "2"},
+		}}
+		testEvalContext.Event.Response.Responses = []types.SurveyItemResponse{
+			{Key: "weekly.G1.Q1", Response: &types.ResponseItem{Key: "rg", Items: []types.ResponseItem{{Key: "mcg", Items: []types.ResponseItem{
+				{Key: "0"},
+				{Key: "3"},
+			}}}}},
+		}
+		ret, err := ExpressionEval(exp, testEvalContext)
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+			return
+		}
+		if !ret.(bool) {
+			t.Errorf("unexpected value: %b", ret)
+		}
+	})
+}
+
 func TestEvalResponseHasKeysAny(t *testing.T) {
 	testEvalContext := EvalContext{
 		Event: types.StudyEvent{